refactor(YALB-639): namespaced second half of molecules
