+++ components/00-tokens/colors/color-pairings.twig
@@ -15,7 +15,7 @@
         cta__href: '#',
       } %}
     </div>
-    {% include "@molecules/social-links/social-links.twig" with {
+    {% include "@molecules/social-links/yds-social-links.twig" with {
       icons: {
         1: {
           name: 'Facebook',
