+++ components/02-molecules/alert/alert.stories.js
@@ -1,5 +1,5 @@
 import alertTwig from './yds-alert.twig';
-import textFieldTwig from '../text/text-field.twig';
+import textFieldTwig from '../text/yds-text-field.twig';
 import ctaTwig from '../../01-atoms/controls/cta/yds-cta.twig';
 
 import alertData from './alert.yml';
