+++ components/02-molecules/molecules.scss
@@ -10,10 +10,10 @@
 @forward './meta/basic-meta/yds-basic-meta';
 @forward './meta/event-meta/yds-event-meta';
 @forward './page-title/yds-page-title';
-@forward './pager/pager';
-@forward './pop-out-image/pop-out-image';
-@forward './pull-quote/pull-quote';
-@forward './social-links/social-links';
-@forward './tabs/tabs';
-@forward './text/text-field';
-@forward './text-with-image/text-with-image';
+@forward './pager/yds-pager';
+@forward './pop-out-image/yds-pop-out-image';
+@forward './pull-quote/yds-pull-quote';
+@forward './social-links/yds-social-links';
+@forward './tabs/yds-tabs';
+@forward './text/yds-text-field';
+@forward './text-with-image/yds-text-with-image';
