+++ components/02-molecules/pop-out-image/_yds-pop-out-image.scss
