+++ components/02-molecules/social-links/yds-social-links.twig
@@ -17,7 +17,7 @@
       {# list-items <li> #}
       {# social-links__item #}
       <li {{ bem('item', [], social_links__baseclass) }}>
-        {% include "@molecules/social-links/_social-link__content.twig" %}
+        {% include "@molecules/social-links/_yds-social-link__content.twig" %}
       </li>
     {% endfor %}
   </ul>
