+++ components/02-molecules/tabs/_yds-tab-content.twig
