+++ components/02-molecules/tabs/_yds-tabs.scss
