+++ components/02-molecules/text-with-image/_yds-text-with-image.scss
