+++ components/03-organisms/layout/two-column/_two-column--example.twig
@@ -1,9 +1,9 @@
 {% extends "@organisms/layout/two-column/two-column.twig" %}
   {% block two_column__primary %}
-    {% include "@molecules/text/text-field.twig" %}
+    {% include "@molecules/text/yds-text-field.twig" %}
   {% endblock %}
   {% block two_column__secondary %}
-    {% include "@molecules/text/text-field.twig" with {
+    {% include "@molecules/text/yds-text-field.twig" with {
       text_field__content: '<h2>Program Information</h2><h3>The Undergraduate Handbook</h3><p>A comprehensive guide for prospective and current Chemistry Majors, with a complete description of requirements and opportunities.</p><p><a href="#">View the handbook online</a></p><h3>Course List</h3><p>A list of courses and a description of the Chemistry undergraduate program, including information on placement exams, laboratory registration, premedical students, and major requirements</p><p><a href="#">Yale College programs of study</a></p>',
     } %}
   {% endblock %}
