+++ components/03-organisms/site-footer/site-footer.twig
@@ -50,7 +50,7 @@
       {# Social #}
       <div {{ bem('social', [], site_footer__base_class) }}>
         {% block site_footer__social %}
-          {% include "@molecules/social-links/social-links.twig" %}
+          {% include "@molecules/social-links/yds-social-links.twig" %}
         {% endblock %}
       </div>
       {# Meta #}
