+++ components/05-page-examples/_intro-content-examples.twig
@@ -13,13 +13,13 @@
     content_image__width: 'max',
   } %}
 {% elseif intro_content == 'pop-out-image' %}
-  {% include "@molecules/pop-out-image/pop-out-image.twig" with {
+  {% include "@molecules/pop-out-image//yds-pop-out-image.twig" with {
     pop_out_image__content: '<p>Dr. Davis’s research group at the Kline Chemistry Laboratory uses experiments at multiple scales – in vitro, single cell, and whole organism – to study fundamental and applied problems at the intersection of chemistry, physics, and biology. They develop new quantitative spectroscopic imaging techniques to elucidate the relationship between function and dynamics of proteins and RNA inside living cells. </p><p>Caitlin Davis obtained her Ph.D. from Emory University in 2015, where she studied protein folding in the laboratory of Dr. Brian Dyer in the Chemistry Department. She completed her postdoctoral training with Dr. Martin Gruebele at the Center for the Physics of Living Cells at the University of Illinois at Urbana-Champaign, where she developed a method for studying protein thermodynamics and kinetics in differentiated tissues of living zebrafish and she developed a mimic of cytoplasm that can be used to reproduce protein behaviors in vitro. She came to Yale as a faculty member in 2020.</p>',
   } %}
 {% elseif intro_content == 'text-with-image' %}
-  {% include "@molecules/text-with-image/text-with-image.twig" %}
+  {% include "@molecules/text-with-image/yds-text-with-image.twig" %}
 {% elseif intro_content == 'text-with-image--highlight' %}
-  {% include "@molecules/text-with-image/text-with-image.twig" with {
+  {% include "@molecules/text-with-image/yds-text-with-image.twig" with {
     text_with_image__width: 'highlight',
   } %}
 {% elseif intro_content == 'collection-featured' %}
