+++ components/05-page-examples/events/_event-collection--example.twig
@@ -40,5 +40,5 @@
       {},
     ]
   } %}
-  {% include "@molecules/pager/pager.twig" %}
+  {% include "@molecules/pager//yds-pager.twig" %}
 {% endif %}
