+++ components/05-page-examples/news/news-article.twig
@@ -14,7 +14,7 @@
     {% include "@molecules/image/yds-content-image.twig" with {
       content_image__width: 'highlight',
     } %}
-    {% include "@molecules/text/text-field.twig" with {
+    {% include "@molecules/text/yds-text-field.twig" with {
       text_field__content: '<p>Caitlin Davis, assistant professor of chemistry, has received a 2021 Scialog Collaborative Innovation Award to pursue research into the structure and function of enzyme filaments that regulate cell metabolism. Understanding cell processes is important as their alteration instigates the growth and spread of disease inside the body.</p><p>The award, from the Research Corporation for Science Advancement (RCSA) and the Gordon and Betty Moore Foundation, is part of RCSA’s Scialog Chemical Machinery of the Cell initiative to advance the fundamental understanding of chemical machinery and reactions in the intact cell. Teams of early career researchers who had not previously collaborated competed for seed funding for high-risk, high-reward projects to pursue novel lines of research.</p>',
     } %}
     {% include "@molecules/image/yds-content-image.twig" with {
@@ -23,16 +23,16 @@
       image__alt: image__alt__1,
       image__src: image__src__1,
     } %}
-    {% include "@molecules/text/text-field.twig" with {
+    {% include "@molecules/text/yds-text-field.twig" with {
       text_field__content: '<p>Davis and her teammate Lars Plate, assistant professor of chemistry and biological sciences at Vanderbilt University, were two of 21 researchers to receive the award. They will work together on their project, <em>“</em><em>Structure-Function of Enzyme Filaments: Regulators of Cell Metabolism in Space and Time.”</em></p>',
     } %}
-    {% include "@molecules/pull-quote/pull-quote.twig" with {
+    {% include "@molecules/pull-quote/yds-pull-quote.twig" with {
       pull_quote__quote: 'Our platform will transform our ability to study metabolic filaments and, coupled with biochemical and biophysical tools, manipulate enzyme structures and metabolic activity in cells.',
     } %}
-    {% include "@molecules/text/text-field.twig" with {
+    {% include "@molecules/text/yds-text-field.twig" with {
       text_field__content: '<p>“Our platform will transform our ability to study metabolic filaments and, coupled with biochemical and biophysical tools, manipulate enzyme structures and metabolic activity in cells,” said Davis. “Unraveling these interactions will improve disease diagnoses and therapies that target metabolic networks in developmental disorders, premature aging, and cancers.</p><p>The overarching challenge of this research is that the relationship between molecular filament structures at the nanoscale and cellular self-assembly at the microscale remains largely unknown. Adding to the difficulty is that experimental tools do not currently exist to characterize filaments inside cells with high structural specificity or to quantify the effect of filamentation on enzyme function.</p><p>“The problem is a lot of imaging research that’s done uses fluorescent proteins or fluorescent molecules to tag the protein that you’re looking at, and anything that you do to perturb the protein can affect what’s going on inside cells,” said Davis. “So, if you add a tag onto it, then you’re changing the environment.”</p><p>This is where the expertise of Davis and Plate come together. The Plate lab will provide proficiency in chemical biology and proteomics methods to map the dynamics of protein interactions during the metabolon assembly, while the Davis lab will contribute the infrared imaging to monitor structural changes in live cells.</p><p>Davis explained by using infrared spectroscopy, which is a label-free approach and therefore minimally invasive, researchers can see what’s going on with proteins without changing anything inside the cell. Lars is designing a very small tag to use in the infrared to attract the protein and prove that it is the protein interest.</p><p>The team hypothesizes that cells form distinct types of filaments, activated and inhibited, with unique structures that fine-tune metabolic activity. What’s new about their approach is they will study protein structures by infrared microscopy rather than fluorescence microscopy. They expect these images to reveal a new localized layer of cellular metabolism that could be used to re-engineer metabolic processes. In the future, these complementary time-resolved imaging and cross-linking approaches could be extended to biological studies of macromolecular assemblies and phase-separated membraneless organelles that have been historically difficult to study.</p>',
     } %}
-    {% include "@molecules/pop-out-image/pop-out-image.twig" with {
+    {% include "@molecules/pop-out-image//yds-pop-out-image.twig" with {
       pop_out_image__content: '<p>Dr. Davis’s research group at the Kline Chemistry Laboratory uses experiments at multiple scales – in vitro, single cell, and whole organism – to study fundamental and applied problems at the intersection of chemistry, physics, and biology. They develop new quantitative spectroscopic imaging techniques to elucidate the relationship between function and dynamics of proteins and RNA inside living cells. </p><p>Caitlin Davis obtained her Ph.D. from Emory University in 2015, where she studied protein folding in the laboratory of Dr. Brian Dyer in the Chemistry Department. She completed her postdoctoral training with Dr. Martin Gruebele at the Center for the Physics of Living Cells at the University of Illinois at Urbana-Champaign, where she developed a method for studying protein thermodynamics and kinetics in differentiated tissues of living zebrafish and she developed a mimic of cytoplasm that can be used to reproduce protein behaviors in vitro. She came to Yale as a faculty member in 2020.</p>',
       pop_out_image__caption: '<p>This is a random image. Doesn\'t it look great?</p>',
       image__srcset: image__srcset__pop_out,
@@ -40,7 +40,7 @@
       image__alt: image__alt__pop_out,
       image__src: image__src__pop_out,
     } %}
-    {% include "@molecules/text/text-field.twig" with {
+    {% include "@molecules/text/yds-text-field.twig" with {
       text_field__content: '<p>Research Corporation for Science Advancement is a private foundation that funds basic research in the physical sciences (astronomy, chemistry, physics, and related fields) at colleges and universities through its Cottrell Scholar and Scialog programs.</p><p>The Gordon and Betty Moore Foundation fosters pathbreaking scientific discovery, environmental conservation, patient care improvements, and preservation of the special character of the San Francisco Bay Area.</p><p>Learn more about the <a href="https://www.caitlindavislab.com/">Davis Lab</a>.</p>',
     } %}
   {% endblock %}
